Move runBasic into utils_test.go and update test since 8.19 cannot upgrade to 9.0
