+++ functionaltests/9_0_test.go
@@ -25,40 +25,3 @@ func TestUpgrade_8_18_to_9_0(t *testing.T) {
 	t.Parallel()
 	runBasicUpgradeSnapshotTest(t, "8.18", "9.0")
 }
-
-func TestUpgrade_8_19_to_9_0(t *testing.T) {
-	t.Parallel()
-	runBasicUpgradeSnapshotTest(t, "8.19", "9.0")
-}
-
-func runBasicUpgradeSnapshotTest(t *testing.T, fromVersionPrefix, toVersionPrefix string) {
-	skipNonActiveVersions(t, toVersionPrefix)
-
-	tt := singleUpgradeTestCase{
-		fromVersion: getLatestSnapshotFor(t, fromVersionPrefix),
-		toVersion:   getLatestSnapshotFor(t, toVersionPrefix),
-		checkPreUpgradeAfterIngest: checkDatastreamWant{
-			Quantity:         8,
-			PreferIlm:        true,
-			DSManagedBy:      managedByILM,
-			IndicesPerDs:     1,
-			IndicesManagedBy: []string{managedByILM},
-		},
-		checkPostUpgradeBeforeIngest: checkDatastreamWant{
-			Quantity:         8,
-			PreferIlm:        true,
-			DSManagedBy:      managedByILM,
-			IndicesPerDs:     1,
-			IndicesManagedBy: []string{managedByILM},
-		},
-		checkPostUpgradeAfterIngest: checkDatastreamWant{
-			Quantity:         8,
-			PreferIlm:        true,
-			DSManagedBy:      managedByILM,
-			IndicesPerDs:     1,
-			IndicesManagedBy: []string{managedByILM},
-		},
-	}
-
-	tt.Run(t)
-}
