+++ functionaltests/9_1_test.go
@@ -1,0 +1,8 @@
+package functionaltests
+
+import "testing"
+
+func TestUpgrade_8_19_to_9_1(t *testing.T) {
+	t.Parallel()
+	runBasicUpgradeSnapshotTest(t, "8.19", "9.1")
+}
